refactor: split packaging code into DEB and RPM modules
  Separate build_package.py into three focused files:
  - deb_package.py: Debian packaging functions
  - rpm_package.py: RPM packaging functions
  - build_package.py: Main orchestration

  Improves code organization and maintainability.
